Move data loading progress bar to utilities
